refactor(ai): restructure into app python module
